+++ src/main/java/dev/ftb/mods/ftbic/block/entity/machine/DiggingBaseBlockEntity.java
@@ -4,6 +4,7 @@ import com.mojang.authlib.GameProfile;
 import dev.ftb.mods.ftbic.FTBICConfig;
 import dev.ftb.mods.ftbic.block.ElectricBlockInstance;
 import dev.ftb.mods.ftbic.block.FTBICBlocks;
+import dev.ftb.mods.ftbic.net.MoveLaserPayload;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.SectionPos;
@@ -16,6 +17,7 @@ import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.GameType;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
@@ -30,6 +32,7 @@ import net.neoforged.neoforge.common.CommonHooks;
 import net.neoforged.neoforge.common.Tags;
 import net.neoforged.neoforge.common.util.FakePlayer;
 import net.neoforged.neoforge.common.util.FakePlayerFactory;
+import net.neoforged.neoforge.network.PacketDistributor;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
@@ -232,7 +235,9 @@ public class DiggingBaseBlockEntity extends BasicMachineBlockEntity {
 				laserX = (float) (offsetX + 1 + col + 0.5);
 				laserZ = (float) (offsetZ + 1 + row + 0.5);
 				laserY = my;
-				level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 3);
+				if (level instanceof ServerLevel server) {
+					PacketDistributor.sendToPlayersTrackingChunk(server, ChunkPos.containing(worldPosition), new MoveLaserPayload(worldPosition, laserX, laserY, laserZ));
+				}
 				digBlock(state, miningPos);
 			}
 		}